Added custom JWTAuthentication
jwt.py in users folder will be used for future pages requiring user authentication.
It does so by taking in a token and decoding it. If the decoded token corresponds to a user in the database, it is authenticated.
An example view is being used and tested in users/view.py called AuthUserAPIView.